moved telnet code to telnet io script
